1010: Extracting DomesticPaymentConsentValidationHelpers and ApiTestUtils classes
https://github.com/SecureApiGateway/SecureApiGateway/issues/1010
