+++ secure-api-gateway-ob-uk-rcs-consent-store/secure-api-gateway-ob-uk-rcs-consent-store-api/src/test/java/com/forgerock/sapi/gateway/rcs/consent/store/api/v3_1_10/ApiTestUtils.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright © 2020-2022 ForgeRock AS (obst@forgerock.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.forgerock.sapi.gateway.rcs.consent.store.api.v3_1_10;
+
+import org.springframework.http.HttpHeaders;
+
+public class ApiTestUtils {
+    public static HttpHeaders createConsentStoreApiRequiredHeaders(String apiClientId) {
+        final HttpHeaders headers = new HttpHeaders();
+        headers.add("x-api-client-id", apiClientId);
+        return headers;
+    }
+}
